Attribute tests moved to Attributes namespace
